Sub: move update_sensor_flags to GCS_Sub.cpp
